requirements.txt
Pillow==10.2.0
requests==2.32.3
